packages/public-service/src/modules/auth/auth.controller.ts
import { Body, Controller, Post, ValidationPipe } from '@nestjs/common';

import { AuthService } from './auth.service';
import { ApiUseTags } from '@nestjs/swagger';
import { AuthCredentialsDto } from './dto/auth-credentials.dto';

@Controller('auth')
@ApiUseTags('Auth')
export class AuthController {
  constructor(private authService: AuthService) {}

  @Post('/login')
  signIn(
    @Body(ValidationPipe) authCredentialsDto: AuthCredentialsDto,
  ): Promise<{ accessToken: string }> {
    return this.authService.signIn(authCredentialsDto);
  }
}
